Extract sign in modal into it's own component
