+++ frontend/src/component/App.tsx
@@ -1,9 +1,5 @@
 import React, { Component } from 'react';
-import {
-  BrowserRouter as Router,
-  Route,
-  Switch
-} from 'react-router-dom';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 import { ReCaptcha } from '../service/Captcha.service';
 import { Home } from './pages/Home';
